DW-2378.
Run with or without security without modification.

Allow the service to run with security or without it based on
supplied arguments, i.e. to enable mutual authorisation there is no need
to redeploy updated code, simply set the appropriate parameters when the
server is started.

To run with security and mutual authorisation enabled the "SECURE"
spring profile must be active and various application properties must be
set, viz:

server.port 8443
server.ssl.key-store resources/keystore.jks
server.ssl.key-store-password changeit
server.ssl.key-store-type JKS
server.ssl.key-alias ca
server.http2.enabled true
server.ssl.trust-store resources/truststore2.jks
server.ssl.trust-store-password changeit
server.ssl.client-auth need

The values given allow the service to run with some self-signed
certificates that have been put in the ‘/resources’ folder.

To run without security the INSECURE spring profile must be active.

Two run configurations have been committed that set these parameters
accordingly, viz.: "DataKeyServiceApplication - secure" and
"DataKeyServiceApplication - insecure".

To hit an endpoint when in secure mode:

curl -sS --insecure -H "Content-type: text/plain" \
    --cert certificate.pem:changeit --key key.pem \
    https://localhost:8443/datakey/actions/decrypt?keyId=123 \
    --data-binary $(cat encrypted.txt)

From the resources directory at the root of the project hierarchy.

In the interests of full disclosure it should be noted that CSRF
protection is disabled in both insecure and secure modes until I can
fully understand how to apply it appropriately.
