+++ .gitignore
@@ -6,7 +6,5 @@ HELP.md
 # Terraform
 ci/terraform.tf
 *.tfvars
-.idea/
 rad/
-*.iml
 out/
