+++ .idea/runConfigurations/DataKeyServiceApplication___insecure.xml
@@ -1,0 +1,11 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="DataKeyServiceApplication - insecure" type="SpringBootApplicationConfigurationType" factoryName="Spring Boot">
+    <module name="data-key-service.main" />
+    <option name="SPRING_BOOT_MAIN_CLASS" value="uk.gov.dwp.dataworks.DataKeyServiceApplication" />
+    <option name="ACTIVE_PROFILES" value="KMS,AWS,INSECURE" />
+    <option name="ALTERNATIVE_JRE_PATH" />
+    <method v="2">
+      <option name="Make" enabled="true" />
+    </method>
+  </configuration>
+</component>
