+++ .idea/runConfigurations/DataKeyServiceApplication___secure.xml
@@ -1,0 +1,58 @@
+<component name="ProjectRunConfigurationManager">
+  <configuration default="false" name="DataKeyServiceApplication - secure" type="SpringBootApplicationConfigurationType" factoryName="Spring Boot">
+    <module name="data-key-service.main" />
+    <option name="SPRING_BOOT_MAIN_CLASS" value="uk.gov.dwp.dataworks.DataKeyServiceApplication" />
+    <option name="ACTIVE_PROFILES" value="KMS,AWS,SECURE" />
+    <option name="ALTERNATIVE_JRE_PATH" />
+    <additionalParameters>
+      <param>
+        <option name="enabled" value="true" />
+        <option name="name" value="server.port" />
+        <option name="value" value="8443" />
+      </param>
+      <param>
+        <option name="enabled" value="true" />
+        <option name="name" value="server.ssl.key-store" />
+        <option name="value" value="resources/keystore.jks" />
+      </param>
+      <param>
+        <option name="enabled" value="true" />
+        <option name="name" value="server.ssl.key-store-password" />
+        <option name="value" value="changeit" />
+      </param>
+      <param>
+        <option name="enabled" value="true" />
+        <option name="name" value="server.ssl.key-store-type" />
+        <option name="value" value="JKS" />
+      </param>
+      <param>
+        <option name="enabled" value="true" />
+        <option name="name" value="server.ssl.key-alias" />
+        <option name="value" value="ca" />
+      </param>
+      <param>
+        <option name="enabled" value="true" />
+        <option name="name" value="server.http2.enabled" />
+        <option name="value" value="true" />
+      </param>
+      <param>
+        <option name="enabled" value="true" />
+        <option name="name" value="server.ssl.trust-store" />
+        <option name="value" value="resources/truststore2.jks" />
+      </param>
+      <param>
+        <option name="enabled" value="true" />
+        <option name="name" value="server.ssl.trust-store-password" />
+        <option name="value" value="changeit" />
+      </param>
+      <param>
+        <option name="enabled" value="true" />
+        <option name="name" value="server.ssl.client-auth" />
+        <option name="value" value="need" />
+      </param>
+    </additionalParameters>
+    <method v="2">
+      <option name="Make" enabled="true" />
+    </method>
+  </configuration>
+</component>
