+++ resources/encrypted.txt
@@ -1,0 +1,1 @@
+AQIDAHiDZm7dsImftTfAGQhiuj+6YBT6kuD9jOcAaAq8C3vEJAF3odArCwqBuySyWi8YeUJ6AAAAbjBsBgkqhkiG9w0BBwagXzBdAgEAMFgGCSqGSIb3DQEHATAeBglghkgBZQMEAS4wEQQMmDLypwFienjhlT6nAgEQgCvAGXA0SkZSIAcWWClf2artrR/80kbavJMugaxSqHgxABNJFi5AG8FrjcXu
