+++ src/main/java/uk/gov/dwp/dataworks/config/InsecureConfiguration.java
@@ -1,0 +1,22 @@
+package uk.gov.dwp.dataworks.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Profile;
+import org.springframework.http.HttpMethod;
+import org.springframework.security.config.annotation.web.builders.HttpSecurity;
+import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+
+@Configuration
+@Profile("INSECURE")
+public class InsecureConfiguration extends WebSecurityConfigurerAdapter {
+
+    @Override
+    protected void configure(HttpSecurity httpSecurity) throws Exception {
+        httpSecurity.csrf().disable()
+                .authorizeRequests()
+                .antMatchers("/")
+                .permitAll()
+                .antMatchers(HttpMethod.POST, "/")
+                .permitAll();
+    }
+}
