+++ src/test/java/uk/gov/dwp/dataworks/integration/HealthCheckIntegrationTests.java
@@ -25,7 +25,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @RunWith(SpringRunner.class)
 @SpringBootTest()
-@ActiveProfiles("IntegrationTest")
+@ActiveProfiles({"IntegrationTest", "INSECURE"})
 @AutoConfigureMockMvc
 public class HealthCheckIntegrationTests {
