+++ src/test/java/uk/gov/dwp/dataworks/integration/PingIntegrationTests.java
@@ -15,7 +15,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 @RunWith(SpringRunner.class)
 @SpringBootTest()
-@ActiveProfiles("IntegrationTest")
+@ActiveProfiles({"IntegrationTest", "INSECURE"})
 @AutoConfigureMockMvc
 public class PingIntegrationTests {
